generator/requirements.txt
# Core template engine
jinja2==3.1.2

# Markdown processing for content
markdown==3.5.1

# YAML configuration files
pyyaml==6.0.1

# EPUB generation
ebooklib==0.18

# HTML parsing for broken link detection
beautifulsoup4==4.12.2

# Image processing for WebP optimization
Pillow==10.1.0

# File watching for live reload functionality
watchdog==3.0.0

# WebSocket server for live reload
websockets==12.0

# Asset minification
htmlmin==0.1.12
rcssmin==1.1.1
rjsmin==1.2.0